Database Cleanup (#2281 rework) #2383
neskk: Pulled the changes from #2281, rebased it and added a WIP
SpawnPoint cleanup routine.
Since SpawnPoint are a bit tricky and are connected to other three
tables (ScanSpawnPoint, SpawnPointDetectionData and ScannedLocation) I
tried to cascade the cleanup so that all relevant information is kept
and only incorrect data is deleted.